Moved Stats components into Stats dir
