+++ src/components/Dashboard/Stats/Chart.jsx
@@ -8,7 +8,7 @@ import {
   Label,
   ResponsiveContainer
 } from "recharts";
-import Title from "./Title";
+import Title from "../Title";
 
 // Generate Sales Data
 function createData(time, amount) {
