+++ src/components/Dashboard/Stats/Deposits.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import Link from "@material-ui/core/Link";
 import { makeStyles } from "@material-ui/core/styles";
 import Typography from "@material-ui/core/Typography";
-import Title from "./Title";
+import Title from "../Title";
 
 function preventDefault(event) {
   event.preventDefault();
