+++ src/components/Dashboard/Stats/Orders.jsx
@@ -6,7 +6,7 @@ import TableBody from "@material-ui/core/TableBody";
 import TableCell from "@material-ui/core/TableCell";
 import TableHead from "@material-ui/core/TableHead";
 import TableRow from "@material-ui/core/TableRow";
-import Title from "./Title";
+import Title from "../Title";
 
 // Generate Order Data
 function createData(id, date, name, shipTo, paymentMethod, amount) {
